+++ examples/coap-proxy/simple-coap-server/virtual-sensor.h
@@ -40,7 +40,7 @@
 #ifndef VIRTUAL_SENSOR_H
 #define VIRTUAL_SENSOR_H
 
-float read_temperature();
-float read_humidity();
+int read_temperature();
+int read_humidity();
 
 #endif
